Fix memory leaks
When a node is removed from the tree but not freed, the application
leaks memory on exit. This cannot be fixed by freeing these references
on tree exit, so it seems that the only way to solve this problem is
by using PackedSceneReferences (which is the accepted convention in
Godot for how to create instances of a scene anyway).